requirements.txt
psycopg2-binary
pgvector
python_dotenv
requests
pymupdf
openai
langchain_text_splitters
tqdm
flask